requirements.txt
openai>=1.45.0
pyyaml
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
requests
PyGithub>=2.0.0
